requirements.txt
langchain-openai
python-dotenv
youtube-transcript-api
chromadb
langchain-chroma
langchain-text-splitters
langchain-community